requirements.txt
aiohttp==3.6.2
apt-clone==0.2.1
apturl==0.5.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.6.0
Brlapi==0.6.6
catfish==1.4.10
certifi==2018.1.18
chardet==3.0.4
command-not-found==0.3
configobj==5.0.6
cupshelpers==1.0
defer==1.0.6
discord.py==1.3.3
httplib2==0.9.2
idna==2.9
idna-ssl==1.1.0
louis==3.5.0
Mako==1.0.7
MarkupSafe==1.0
menulibre==2.2.1
mugshot==0.4.2
multidict==4.7.6
netifaces==0.10.4
onboard==1.4.1
pexpect==4.2.1
Pillow==5.1.0
psutil==5.4.2
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
pygobject==3.26.1
PyICU==1.9.8
pyparted==3.11.1
python-apt==1.6.5+ubuntu0.3
python-debian==0.1.32
python-dotenv==0.13.0
python-xapp==1.8.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
sessioninstaller==0.0.0
setproctitle==1.1.10
six==1.11.0
system-service==0.3
systemd-python==234
typing-extensions==3.7.4.2
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.22
websockets==8.1
xkit==0.0.0
yarl==1.4.2
